menu_project/requirements.txt
Django==4.1.7
djangorestframework==3.14.0
psycopg2-binary==2.9.5
python-dotenv==1.0.0
uvicorn==0.21.0
